+++ packages/uniforms-bridge-simple-schema-2/__tests__/SimpleSchema2Bridge.ts
@@ -332,8 +332,13 @@ describe('SimpleSchema2Bridge', () => {
 
   describe('#getValidator', () => {
     it('calls correct validator', () => {
-      expect(() => bridge.getValidator()({})).toThrow();
-      expect(() => bridge.getValidator({})({})).toThrow();
+      const schema = new SimpleSchema({ x: { type: Number } });
+      const bridge = new SimpleSchema2Bridge(schema);
+
+      expect(bridge.getValidator()({})).not.toEqual(null);
+      expect(bridge.getValidator({})({})).not.toEqual(null);
+      expect(bridge.getValidator()({ x: 1 })).toEqual(null);
+      expect(bridge.getValidator({})({ x: 1 })).toEqual(null);
     });
   });
 });