+++ packages/uniforms/src/BaseForm.tsx
@@ -16,7 +16,7 @@ export type BaseFormProps<Model> = {
   autosave: boolean;
   autosaveDelay: number;
   disabled?: boolean;
-  error?: any;
+  error: any;
   id?: string;
   label: boolean;
   model: DeepPartial<Model>;
@@ -26,9 +26,7 @@ export type BaseFormProps<Model> = {
   ) => DeepPartial<Model>;
   noValidate: boolean;
   onChange?(key: string, value: any): void;
-  onSubmit?(model: DeepPartial<Model>): any;
-  onSubmitFailure?(result: any): void;
-  onSubmitSuccess?(result: any): void;
+  onSubmit(model: DeepPartial<Model>): any;
   placeholder?: boolean;
   schema: any;
   showInlineError?: boolean;
@@ -51,9 +49,11 @@ export class BaseForm<
   static defaultProps = {
     autosave: false,
     autosaveDelay: 0,
+    error: null,
     label: true,
     model: Object.create(null),
     noValidate: true,
+    onSubmit() {},
   };
 
   constructor(props: Props) {
@@ -250,24 +250,15 @@ export class BaseForm<
       event.stopPropagation();
     }
 
-    const result =
-      this.props.onSubmit && this.props.onSubmit(this.getModel('submit'));
-
-    // Set the `submitting` state only if onSubmit is async so we don't cause an unnecessary re-render
-    let submitting: Promise<any>;
-    if (isPromiseLike(result)) {
-      this.setState({ submitting: true });
-      submitting = result.finally(() => {
-        this.setState({ submitting: false });
-      });
-    } else {
-      submitting = Promise.resolve(result);
+    const result = this.props.onSubmit(this.getModel('submit'));
+    if (!(result instanceof Promise)) {
+      return Promise.resolve(result);
     }
 
-    return submitting.then(
-      this.props.onSubmitSuccess,
-      this.props.onSubmitFailure,
-    );
+    this.setState({ submitting: true });
+    return result.finally(() => {
+      this.setState({ submitting: false });
+    });
   }
 
   render() {
@@ -277,8 +268,4 @@ export class BaseForm<
       </context.Provider>
     );
   }
-}
-
-function isPromiseLike(value: any): value is Promise<any> {
-  return !!value && isFunction(value.then);
 }